refactor(flightdeck-core): extract pane-registry helpers out of daemon/loop.ts
W5 reviewer-structure follow-up B4: daemon/loop.ts grew 581 -> 828
lines after the W5 sweep. The pane-registry adapters (paneRegistryArgs,
paneRegistryIssueForPane, extractFlag, resolveMeta, paneRegistryRows,
resolvePaneTargetForEntry, entryKindForPane, listTrackedEntriesForReconcile)
are pure shell-out + JSON parse helpers that don't reference any loop
state. Moving them to a new daemon/pane-registry.ts module trims
loop.ts back closer to its original size and gives the helpers a
proper home.

No behavioral change: loop.ts now imports the helpers and re-exports
listTrackedEntriesForReconcile so reconcile.ts and any sibling that
imported it from loop.ts still works without modification.
